+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - `XIT ACT`: Fix the CX Buy action getting stuck when there are no orders in the order book and the "buy partial" toggle is on
 - `XIT ACT`: Fix opening a run tile for packages with configurable MTRA and no origins/destinations available
 - `XIT ACT`: Fix MTRA action getting stuck when there is no space in the destination inventory
+- `XIT ACT`: Fix MTRA action error when the material is not present in origin inventory
 - `XIT SET`: Fix financial data point deletion targeting the wrong data point
 
 ## 25.4.12